requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
numpy>=2.1,<3
neo4j==5.23.1
pydantic==2.8.2
requests==2.32.3
ollama==0.3.1
